+++ CHANGELOG.md
@@ -4,13 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.4-alpha] - XXXX-XX-XX
+## [0.0.4-alpha] - 2021-12-05
  
 ### Added
+- Implementation of AVFrame
 
 ### Changed
 
 ### Fixed
+- Errors inside AVPacketImpl now logs the error before throwing
+- Check error on IAVPacket::redToPacket()
+- Fixed AVPacket::getRawData() for non-referenced packets
 
 ## [0.0.3-alpha] - 2021-12-04